leave-system/requirements.txt
# Michaelhouse Leave System - Python Dependencies

# Core Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Email
secure-smtplib==0.1.1

# Environment
python-dotenv==1.0.0

# Date/Time
python-dateutil==2.8.2

# Development
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.1
flake8==6.1.0

# Production Server
gunicorn==21.2.0
